+++ Editor/CodeBindData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace CodeBind.Editor
 {
-    internal sealed class CodeBindData : IComparable<CodeBindData>
+    public sealed class CodeBindData : IComparable<CodeBindData>
     {
         public string BindName
         {
@@ -27,10 +27,10 @@ namespace CodeBind.Editor
 
         public CodeBindData(string bindName, Type bindType, string bindPrefix, Transform bindTransform)
         {
-            BindName = bindName;
-            BindType = bindType;
-            BindPrefix = bindPrefix;
-            BindTransform = bindTransform;
+            this.BindName = bindName;
+            this.BindType = bindType;
+            this.BindPrefix = bindPrefix;
+            this.BindTransform = bindTransform;
         }
 
         public int CompareTo(CodeBindData other)