+++ Editor/DefaultCodeBindCustomizer.cs
@@ -9,27 +9,17 @@ namespace CodeBind.Editor
     {
         public int Priority => 0;
 
-        public string GetFieldName(string bindName, string bindPrefix)
+        public string GetFieldName(string name)
         {
-            return $"m_{bindName}{bindPrefix}";
+            return $"m_{name}";
         }
 
-        public string GetPropertyName(string bindName, string bindPrefix)
+        public string GetPropertyName(string name)
         {
-            return $"{bindName}{bindPrefix}";
+            return name;
         }
 
-        public string GetArrayFieldName(string bindName, string bindPrefix)
-        {
-            return $"m_{bindName}{bindPrefix}Array";
-        }
-
-        public string GetArrayPropertyName(string bindName, string bindPrefix)
-        {
-            return $"{bindName}{bindPrefix}Array";
-        }
-
-        public string GenerateExtraCode(string nameSpace, string className, IReadOnlyList<CodeBindMemberInfo> members, IReadOnlyList<CodeBindArrayMemberInfo> arrayMembers, string indentation)
+        public string GenerateExtraCode(string nameSpace, string className, List<CodeBindData> bindDatas, SortedDictionary<string, List<CodeBindData>> bindArrayDataDict, string indentation)
         {
             return string.Empty;
         }